# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type StoreCreditCheckoutInputShippingType = typeof StoreCreditCheckoutInputShippingType[keyof typeof StoreCreditCheckoutInputShippingType];


export const StoreCreditCheckoutInputShippingType = {
  standard: 'standard',
  express: 'express',
  'us-warehouse': 'us-warehouse',
} as const;

export type StoreCreditCheckoutItemFromWarehouse = typeof StoreCreditCheckoutItemFromWarehouse[keyof typeof StoreCreditCheckoutItemFromWarehouse];


export const StoreCreditCheckoutItemFromWarehouse = {
  us: 'us',
} as const;

export interface StoreCreditCheckoutItem {
  /**
     * @minLength 1
     * @maxLength 160
     */
  name: string;
  /**
     * @minLength 1
     * @maxLength 80
     */
  dose: string;
  /**
     * @minimum 1
     * @maximum 50
     */
  quantity: number;
  fromWarehouse?: StoreCreditCheckoutItemFromWarehouse;
  /** @maxLength 80 */
  noteLabel?: string;
}

export interface StoreCreditCheckoutForm {
  /**
     * @minLength 1
     * @maxLength 100
     */
  firstName: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  lastName: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  country: string;
  /**
     * @minLength 1
     * @maxLength 250
     */
  address: string;
  /** @maxLength 250 */
  address2: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  city: string;
  /** @maxLength 100 */
  state: string;
  /**
     * @minLength 1
     * @maxLength 40
     */
  postalCode: string;
  /**
     * @minLength 1
     * @maxLength 40
     */
  phone: string;
  /** @maxLength 40 */
  taxId: string;
}

export const StoreCreditPurchaserAttestationValue = {
  over21AndResearchUseOnly: true,
  qualifiedResearcherOrLicensedProfessional: true,
  noHumanOrAnimalUse: true,
  policiesAccepted: true,
} as const;
export type StoreCreditPurchaserAttestation = typeof StoreCreditPurchaserAttestationValue;

export interface StoreCreditCheckoutInput {
  /** @pattern ^INV-[A-Fa-f0-9]{32}$ */
  orderId: string;
  /**
     * @minItems 1
     * @maxItems 100
     */
  items: StoreCreditCheckoutItem[];
  checkoutForm: StoreCreditCheckoutForm;
  shippingType: StoreCreditCheckoutInputShippingType;
  /** @maxLength 40 */
  paymentMethod: string;
  /** @maxLength 80 */
  promoCode: string;
  ownerFreeShipping: boolean;
  /** @maxLength 80 */
  affiliateCode: string;
  affiliateDiscountDisabled: boolean;
  /** @exclusiveMinimum 0 */
  storeCreditUsed: number;
  /** @maxLength 2000 */
  orderNotes: string;
  purchaserAttestation: StoreCreditPurchaserAttestation;
}

export type StoreCreditCheckoutResultOrder = { [key: string]: unknown };

export interface StoreCreditCheckoutResult {
  ok: true;
  order: StoreCreditCheckoutResultOrder;
  /** @minimum 0 */
  balance: number;
  replayed: boolean;
}

export interface CheckoutError {
  ok: false;
  /** @minLength 1 */
  error: string;
}

export interface PublicPromoCode {
  /**
     * @minLength 1
     * @maxLength 80
     */
  code: string;
  /**
     * @maximum 1
     * @exclusiveMinimum 0
     */
  rate: number;
}

export interface PublicPromoCodeLookupResult {
  ok: true;
  promo: PublicPromoCode | null;
}

export interface PublicPromoCodeError {
  ok: false;
  /** @minLength 1 */
  error: string;
}

export interface SupportAttachmentUploadInput {
  /**
     * @minLength 1
     * @maxLength 255
     */
  fileName: string;
  /**
     * @minLength 1
     * @maxLength 120
     */
  contentType: string;
  /**
     * @minimum 1
     * @maximum 104857600
     */
  size: number;
}

export interface SupportAttachmentUploadResult {
  ok: true;
  /**
     * @minLength 1
     * @maxLength 200
     */
  path: string;
  uploadUrl: string;
  /**
     * @minLength 1
     * @maxLength 120
     */
  contentType: string;
}

export type SupportAttachmentAccessItemField = typeof SupportAttachmentAccessItemField[keyof typeof SupportAttachmentAccessItemField];


export const SupportAttachmentAccessItemField = {
  message: 'message',
  admin_reply: 'admin_reply',
} as const;

export interface SupportAttachmentAccessItem {
  /**
     * @minLength 1
     * @maxLength 2048
     */
  path: string;
  /** @minimum 1 */
  messageId?: number;
  field?: SupportAttachmentAccessItemField;
}

export interface SupportAttachmentAccessInput {
  /**
     * @minItems 1
     * @maxItems 100
     */
  attachments: SupportAttachmentAccessItem[];
}

export interface SupportAttachmentAccessUrl {
  /**
     * @minLength 1
     * @maxLength 2048
     */
  reference: string;
  url: string;
}

export interface SupportAttachmentAccessResult {
  ok: true;
  /** @maxItems 100 */
  attachments: SupportAttachmentAccessUrl[];
}

export interface SupportAttachmentError {
  ok: false;
  /** @minLength 1 */
  error: string;
}

export type LookupPublicPromoCodeParams = {
/**
 * @minLength 1
 * @maxLength 80
 * @pattern ^[A-Za-z0-9_-]+$
 */
code: string;
};