# lib/api-zod/src/generated/types/checkoutError.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CheckoutError {
  ok: false;
  /** @minLength 1 */
  error: string;
}