# lib/api-zod/src/generated/types/storeCreditCheckoutForm.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface StoreCreditCheckoutForm {
  /**
     * @minLength 1
     * @maxLength 100
     */
  firstName: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  lastName: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  country: string;
  /**
     * @minLength 1
     * @maxLength 250
     */
  address: string;
  /** @maxLength 250 */
  address2: string;
  /**
     * @minLength 1
     * @maxLength 100
     */
  city: string;
  /** @maxLength 100 */
  state: string;
  /**
     * @minLength 1
     * @maxLength 40
     */
  postalCode: string;
  /**
     * @minLength 1
     * @maxLength 40
     */
  phone: string;
  /** @maxLength 40 */
  taxId: string;
}