# lib/api-zod/src/generated/types/supportAttachmentAccessInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SupportAttachmentAccessItem } from './supportAttachmentAccessItem';

export interface SupportAttachmentAccessInput {
  /**
     * @minItems 1
     * @maxItems 100
     */
  attachments: SupportAttachmentAccessItem[];
}