# lib/api-zod/src/generated/types/supportAttachmentError.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SupportAttachmentError {
  ok: false;
  /** @minLength 1 */
  error: string;
}